Bug 916863: Added Goldfish NFC device
This patch adds an NCI-compatible NFC device to the Goldfish
emulator. The NFC Communication Interface (NCI) is a standardized
protocol for communicating with NFC chipsets. The protocol is
generic enough to be vendor and chipset independend and can be
implemented by devices with NFC-like characteristics.

The NFC device in the Goldfish emulator handles NCI messages, and
responds accordingly. This will allow for simulating basic
functionality of an NFC chipset. Some HCI and proprietary commands
are also supported to allow for use of the BCM2079x driver in the
guest system.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>