Bug 1000935: Added command 'llcp connect' to emulator console
The emulator console now accepts commands for sending LLCP CONNECT
PDUs to the guest system. This feature is not for regular users or
test cases, but only for testing the internals of the emulator!

A possible command looks like

  nfc llcp connect 4 4

where the trailing numbers are the indices of the DSAP and SSAP
Service Access Points. In this example the command would connect
the SAPs for SNEP, which are used for transfering NDEF messages.
SAP numbers are standardized by the NFC Forum.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>