Bug 1000935: Added infrastructure for sending LLCP PDUs in NCI packets
LLCP peers mutually pass send permission, and LLCP requires PDUs
to be sent at specific points when the peer has the permission to
do so. If no PDUs are available for transmission, a SYMM PDU has
to be sent for passing permissions to the other peer.

This patch adds the necessary infrastructure to inject LLCP PDUs
into the stream of SYMM PDUs. If the user creates a new LLCP PDU
it is either

  - send immediatelly if the peer has permisison to do so, or
  - queued up to be sent when permission has been received.

Just as before, if no PDUs are queued up, a SYMM PDU will be sent
after a timeout has been reached. Sending a PDU also passes send
permission to the peer.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>